+++ CHANGELOG.md
@@ -103,7 +103,7 @@ New deprecation(s):
 ### Other
 
 - **General**: Use informer cache for ReplicaSet lookups in GetCurrentReplicas to reduce API server load ([#7466](https://github.com/kedacore/keda/pull/7466))
-- TODO ([#XXX](https://github.com/kedacore/keda/issues/XXX))
+- **Kafa Scaler**: Refactor Kafka Scaler ([#7528](https://github.com/kedacore/keda/pull/7528))
 
 ## v2.19.0